Move transcript styles to separate file
