+++ src/components/styles/ConversationEnded.styles.ts
@@ -19,16 +19,3 @@ export const textStyles: TextStyleProps = {
     marginBottom: "space50",
     marginTop: "space50"
 };
-
-export const buttonStyles: BoxStyleProps = {
-    display: "flex",
-    alignItems: "center",
-    minWidth: "170px"
-};
-
-export const progressStyles: BoxStyleProps = {
-    display: "flex",
-    flexDirection: "column",
-    marginLeft: "space60",
-    alignItems: "flex-start"
-};
